ao_sdl: time out condwaits in playback thread
It's been reported that switching from/to XRoar under Windows causes it
to lock up now.  Without a test machine, the most likely reason seems to
be the audio threads (Windows can restart it, so the thread interaction
is not consistent).

Attempt to fix by allowing each wait to timeout.