Fix AltGr key in windows
In Windows, AltGr presents as Control+Right Alt, so have to ignore Control as
shortcut key if Right Alt pressed.

ToUnicode() also needs side-agnostic keys 'pressed' in supplied state to
recognise them.

ToUnicode() acts as a continuous input method, so one call can affect the
result of the next if a dead key is pressed.  Call it twice so that dead keys
are resolved to their symbol.