mooh: Bank $3F can be mapped at $E000
To allow use of the last bank ($3F), and not waste its 8 KB,
we are for the uppermost $E000-FEFF segment not using $3F as
a "see-through" value (ignore MOOH memory), but this value
will actually map bank $3F here.

This means that there will always be a bank mapped in this
location, as long as the MMU is enabled. The value $3F
remains the default for the corresponding DAT registers,
so bank $3F is therefore by default mapped at $E000.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>