remove redundant tests from configure.ac
Mostly because the results of the tests were not used, but in many cases
also because they duplicate C99 requirements, and C99 already required
to build.

A couple of possible exceptions: malloc() and realloc() might not be GNU
libc compatible under MinGW, but this shouldn't matter, as I never try
to allocate zero bytes.