vo_sdl2: complete rewrite of video initialisation
I think the idea of recreating the window every time it changed size was
carried forward from the SDL1 code.  However, pretty sure we do need to
recreate the renderer.

Setting fullscreen now just uses the SDL call, and the window flag is
checked when the resize event is received.