new approach to SAM VDG counter
Determining how SAM mode changes affect the internal video address
counter.  As it stands, this is pretty much the framework I had a
few years back when I first noticed that the addresses could "jump"
if changed mid-scanline.

However, Stewart Orchard is investigating it from an EE perspective and
has already come up with some insights (not represented here yet):

1. Transitions between dividers can "glitch" - most of these are
deterministic and appear as though the input to a divider were briefly
low, or connected to the output of some other stage.

2. Divider units may be shared between X and Y - there are no modes
where neither X nor Y are in "passthrough" (÷1).