configure: change X11 and Objective C testing
Use the results of AC_PATH_XTRA for X11.

I can't seem to make Objective C compiler detection optional, and if one
is detected - whether or not it's needed - then it gets used for the
final link, using a different set of linker flags.  This is a problem if
some compiler flags (like -flto) are passed to the linker.

To fix this, if Mac OS X Carbon - the only thing that needs Objective C
- is not detected, copy CC into OBJC and CFLAGS into OBJCFLAGS.