keyboard: enhancements to untranslated mode
Matrix points contain a third field which is another scancode to press
first.

New scancode DSCAN_BACKSPACE mapped to Backspace and Delete in UI code
so that it can be treated differently to DSCAN_LEFT.

DSCAN_BACKSPACE mapped differently to DSCAN_LEFT for MC-10.  This way
cursor keys can map to W,Z,A,S without Control pressed (useful for
games) while still allowing Backspace to be mapped to Control+A.