wd279x: write sync bytes before DAM in WRITE SECTOR
This is actually unclear in the datasheet.  Information points to the
DAM initialising the CRC generator to the result of 3 x $A1 sync bytes,
meaning their actual presence may not be required.

However, it doesn't seem sensible, and it could be that in the context
of WRITE SECTOR, the description of "write address mark" could
implicitly require writing those sync bytes.

Reported by Michael Furman.