gtk3: add ComboBoxText with Value
Uses ComboBoxText but creates some extra metadata (provided to "changed"
callback as user_data) to associate a value with entries that isn't just
its index into the list.

Allocated metadata is kept in a list in the UI to be freed on exit.

Debatable whether this is preferable to jumping through the hoops to
have a normal ComboBox with values stored in the list store.  But that
did look really complicated...