ui: add ui_msg_adjust_value_range()
UI message receivers that apply configuration take simple integer
values.  This function applies adjustment and range checking to the
value in a UI message.

UI_PREV and UI_NEXT subtracts or adds 1 to 'cur'.  If
UI_ADJUST_FLAG_CYCLE is set, the result is cycled through 'min' to 'max'
and vice-versa.  Otherwise the result is clamped.

UI_AUTO sets value to 'dfl'.

Note that if the range includes the special command values, i.e. if
'min' is negative, then commands will not take effect.

The new value is returned, and the 'value' field in the UI message is
updated.  If UI_ADJUST_FLAG_KEEP_AUTO is set, the value of UI_AUTO is
kept in the message, even though the returned value differs, allowing an
"automatic" selection to be defined.